Add typings for API
